# config/swagger.yaml
openapi: 3.0.1
info:
  title: npsstore API
  description: This is a Backend API of NBSStore
  version: 1.0.0
  termsOfService: http://quocbaoit.com/terms/
  contact:
    email: NBSstoreProject@quocbaoit.com
...
paths:
...
  /api/v1/producer: 
    post: 
      description: Create a new producer
      summary: Create a new producer
      tags:
        - Producer
      security:
        - bearerAuth: []
      requestBody:
        description: Payload to create a new producer
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerForm'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producer'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/producer/all: 
    get:
      summary:  User get all categories of store
      description: User get all categories of store
      tags:
        - Producer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
  /api/v1/producer/{id}: 
    put:
      summary: Update existed producer
      tags:
        - Producer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Producer ID
          required: true
          schema:
            type: string
      requestBody:
        description: Payload to create a new producer
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerForm'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producer'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete existed producer
      description: Delete existed producer
      security:
        - bearerAuth: []
      tags:
        - Producer
      parameters:
        - name: id
          in: path
          description: Producer ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: 
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/producer/{id}/disable:
    patch:
      summary: Disable a producer
      description: Disable a producer
      security:
        - bearerAuth: []
      tags:
        - Producer
      parameters:
        - name: id
          in: path
          description: Producer ID
          required: true
          schema:
            type: string
      requestBody:
        description: isDisabled value
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isDisabled:
                  type: boolean
                  description: Is disabled or not
                  nullable: false
                  default: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/producer/{id}/restore:
    patch:
      summary: Restore disabled producer
      description: Restore disabled producer
      security:
        - bearerAuth: []
      tags:
        - Producer
      parameters:
        - name: id
          in: path
          description: Producer ID
          required: true
          schema:
            type: string
      requestBody:
        description: isDisabled value
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isDisabled:
                  type: boolean
                  description: Is disabled or not
                  nullable: false
                  default: false
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Producer'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: error message
        stack:
          type: string
          description: error message description
...
    Timestamp:
      type: object
      properties:
        createAt:
          type: string
          description: The date and time when this record is created
          format: date-time
        updatedAt:
          type: string
          description: The date and time when this record is updated
          format: date-time
...
    Category:
      type: object
      properties:
        _id: 
          type: string
          description: Category Id
        name:
          type: string
          description: Category name
        slug:
          type: string
          description: Category slug
        status:
          type: boolean
          description: Category status
          default: true    
        timestamp:
          $ref: '#/components/schemas/Timestamp' 
        createdBy:
          type: string
          description: User object id who created this record
        updatedBy:
          type: string
          description: User object id who updated this record
        isDisabled:
          type: boolean
          description: is Disabled or not
    ProducerForm:
      type: object
      properties:
        name:
          type: string
          description: Producer name
        code:
          type: string
          description: Producer code
        keyword:
          type: boolean
          description: Producer keyword
          default: true 
        status:
          type: boolean
          description: Producer status
          default: true
    Producer:
      type: object
      properties:
        name:
          type: string
          description: Producer name
        code:
          type: string
          description: Producer code
        keyword:
          type: boolean
          description: Producer keyword
          default: true 
        status:
          type: boolean
          description: Producer status
          default: true
        timestamp:
          $ref: '#/components/schemas/Timestamp' 
        createdBy:
          type: string
          description: User object id who created this record
        updatedBy:
          type: string
          description: User object id who updated this record
        isDisabled:
          type: boolean
          description: is Disabled or not
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: apiKey
      description: "<i>JWT Authorization header using the Bearer scheme.</i> <b>Example</b>: <i><u>Bearer 12345abcdef</u></i>"
      in: header          # can be "header", "query" or "cookie"
      name: Authorization # name of the header, query parameter or cookie
# 2) Apply the security globally to all operations
security:
  - bearerAuth: [] # use the same name as above